Use uint32_t for array_id

+++ native-sql-engine/cpp/src/codegen/arrow_compute/ext/array_item_index.h
@@ -24,13 +24,13 @@ namespace arrowcompute {
 namespace extra {
 struct ArrayItemIndex {
   uint16_t id = 0;
-  uint16_t array_id = 0;
+  uint32_t array_id = 0;
   ArrayItemIndex() : array_id(0), id(0) {}
   ArrayItemIndex(uint16_t array_id, uint16_t id) : array_id(array_id), id(id) {}
 };
 struct ArrayItemIndexS {
   uint16_t id = 0;
-  uint16_t array_id = 0;
+  uint32_t array_id = 0;
   ArrayItemIndexS() : array_id(0), id(0) {}
   ArrayItemIndexS(uint16_t array_id, uint16_t id) : array_id(array_id), id(id) {}
 };
